requirements.txt
asgiref==3.2.7
celery==4.4.5
dj-database-url==0.5.0
Django==3.0.7
djangorestframework==3.11.0
django-webpack-loader==0.7.0
psycopg2-binary==2.8.5
python-decouple==3.3
pytz==2020.1
redis==3.5.3
requests==2.24.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
whitenoise==5.1.0
